requirements.txt
asgiref==3.8.1
cloudinary==1.36.0
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==4.2
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
sqlparse==0.5.2
tzdata==2024.2
whitenoise==6.4.0
